Simplify CustomizedRating
