+++ docs/src/pages/components/rating/CustomizedRating.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
-import PropTypes from 'prop-types';
-import { withStyles } from '@material-ui/core/styles';
-import Rating from '@material-ui/core/Rating';
+import { makeStyles, createStyles, withStyles } from '@material-ui/core/styles';
+import Rating, { IconContainerProps } from '@material-ui/core/Rating';
 import FavoriteIcon from '@material-ui/icons/Favorite';
 import FavoriteBorderIcon from '@material-ui/icons/FavoriteBorder';
 import SentimentVeryDissatisfiedIcon from '@material-ui/icons/SentimentVeryDissatisfied';
@@ -10,7 +9,16 @@ import SentimentSatisfiedIcon from '@material-ui/icons/SentimentSatisfied';
 import SentimentSatisfiedAltIcon from '@material-ui/icons/SentimentSatisfiedAltOutlined';
 import SentimentVerySatisfiedIcon from '@material-ui/icons/SentimentVerySatisfied';
 import Typography from '@material-ui/core/Typography';
-import Box from '@material-ui/core/Box';
+
+const useStyles = makeStyles((theme) =>
+  createStyles({
+    root: {
+      '& > legend': {
+        marginTop: theme.spacing(2),
+      },
+    },
+  }),
+);
 
 const StyledRating = withStyles({
   iconFilled: {
@@ -21,7 +29,12 @@ const StyledRating = withStyles({
   },
 })(Rating);
 
-const customIcons = {
+const customIcons: {
+  [index: string]: {
+    icon: React.ReactElement;
+    label: string;
+  };
+} = {
   1: {
     icon: <SentimentVeryDissatisfiedIcon />,
     label: 'Very Dissatisfied',
@@ -44,42 +57,36 @@ const customIcons = {
   },
 };
 
-function IconContainer(props) {
+function IconContainer(props: IconContainerProps) {
   const { value, ...other } = props;
   return <span {...other}>{customIcons[value].icon}</span>;
 }
 
-IconContainer.propTypes = {
-  value: PropTypes.number.isRequired,
-};
+export default function CustomizedRating() {
+  const classes = useStyles();
 
-export default function CustomizedRatings() {
   return (
-    <div>
-      <Box component="fieldset" mb={3} borderColor="transparent">
+    <div className={classes.root}>
         <Typography component="legend">Custom icon and color</Typography>
         <StyledRating
           name="customized-color"
           defaultValue={2}
-          getLabelText={(value) => `${value} Heart${value !== 1 ? 's' : ''}`}
+          getLabelText={(value: number) =>
+            `${value} Heart${value !== 1 ? 's' : ''}`
+          }
           precision={0.5}
           icon={<FavoriteIcon fontSize="inherit" />}
           emptyIcon={<FavoriteBorderIcon fontSize="inherit" />}
         />
-      </Box>
-      <Box component="fieldset" mb={3} borderColor="transparent">
         <Typography component="legend">10 stars</Typography>
         <Rating name="customized-10" defaultValue={2} max={10} />
-      </Box>
-      <Box component="fieldset" mb={3} borderColor="transparent">
         <Typography component="legend">Custom icon set</Typography>
         <Rating
           name="customized-icons"
           defaultValue={2}
-          getLabelText={(value) => customIcons[value].label}
+          getLabelText={(value: number) => customIcons[value].label}
           IconContainerComponent={IconContainer}
         />
-      </Box>
     </div>
   );
 }
